+++ src/scenes/Preloader.ts
@@ -16,6 +16,6 @@ export default class Preloader extends Phaser.Scene {
   }
 
   create() {
-    this.scene.start('game')
+    this.scene.start('game', { level: 1 })
   }
 }
